relocates validate_members fn
