+++ tesseract-server/src/handlers/util.rs
@@ -14,6 +14,10 @@ use tesseract_core::schema::metadata::SourceMetadata;
 
 use crate::app::AppState;
 
+use failure::{bail, format_err, Error};
+use tesseract_core::names::Cut;
+use crate::logic_layer::CubeCache;
+
 
 pub(crate) fn format_to_content_type(format_type: &FormatType) -> ContentType {
     match format_type {
@@ -133,4 +137,19 @@ macro_rules! ok_or_404 {
             }
         }
     };
+}
+
+
+pub fn validate_members(cuts: &[Cut], cube_cache: &CubeCache) -> Result<(), Error> {
+    for cut in cuts {
+        // get level cache
+        let member_cache = cube_cache.members_for_level(&cut.level_name)
+            .ok_or_else(|| format_err!("Level not found in cache"))?;
+        for member in &cut.members {
+            if !member_cache.contains(member) {
+                bail!("Cut member not found");
+            }
+        }
+    }
+    Ok(())
 }
